Fix numerical errors in saving chk
I need to replace conjugate transpose by inv for a in-principle unitary
matrix (eigenvectors of a Hermitian matrix), to avoid numerical issues.
However, this is still causing error when I try to restore the
Hamiltonian using the eigenvalues and eigenvectors.
In the next commit, I will fix this by enforcing Hermiticity when
constructing Hamiltonian.